# Optimize thread_worker