--- a/src/v/redpanda/application.cc
+++ b/src/v/redpanda/application.cc
@@ -2132,7 +2132,7 @@ void application::start_runtime_services(
             });
       });
 
-    thread_worker->start().get();
+    thread_worker->start({.name = "worker"}).get();
 
     // single instance
     node_status_backend.invoke_on_all(&cluster::node_status_backend::start)