--- a/src/v/ssx/fwd.h
+++ b/src/v/ssx/fwd.h
@@ -13,6 +13,6 @@
 
 namespace ssx {
 
-class thread_worker;
+class singleton_thread_worker;
 
 } // namespace ssx