@rockwotj 2023-08-30
> I wonder if it makes sense to add an abort_source, or fail pending requests on stop()?

I am not really sure how to do this correctly, as I don't think seastar classes are designed to support work outside of seastar, unless there is a different technique you're suggesting.

Same goes for failing pending stuff on `stop`, I am not sure how to tell the alien thread that it needs to stop, maybe with an atomic boolean it checks before every task? Thoughts?