@rockwotj 2023-08-31
> LGTM, let's keep an eye on the queue length at shutdown.

I added a followup task for metrics - Also the queue is limited to 128 items for now.